requirements.txt
#asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
#sqlparse==0.5.3
django-redis==5.4.0
django-environ==0.9.0

pillow==11.0.0

uvicorn==0.35.0

channels==4.1.0
channels-redis==4.2.0
whitenoise==6.8.2

psycopg2==2.9.4
